Fixing missing messages from some HTTP exceptions
Some error responses aren't displaying error messages.  For example,
submitting a request with a requested_for value of a username not in
email form (ie, admin instead of admin@myvsphere.local) would return an
error, but not display its output.  It turns out that this error (and
others, I imagine) doesn't contain a systemMessage field.  This change
checks for a systemMessage first, but if one doesn't exist it will use
the message field instead. Not sure if this will cover all possible
error cases, but it at least covers more than we do now.

Also cleaned up rubocop offenses.

Signed-off-by: Nolan Davidson <ndavidson@chef.io>